[KAIZEN-0] refaktoreret oppsett til å samsvare med draft-appen for lettere oppsett av tester
