+++ .gitignore
@@ -14,3 +14,5 @@ src/generated
 src/main/resources/webapp/
 testdb.mv.db
 testdb.trace.db
+
+logs-backup/
