+++ src/main/kotlin/no/nav/modiapersonoversikt/Main.kt
@@ -3,22 +3,19 @@ package no.nav.modiapersonoversikt
 import no.nav.modiapersonoversikt.config.Configuration
 import no.nav.modiapersonoversikt.config.DataSourceConfiguration
 import no.nav.modiapersonoversikt.infrastructure.HttpServer
+import org.slf4j.LoggerFactory
 
-fun runLocally(useAuthentication: Boolean) {
+val log = LoggerFactory.getLogger("modiapersonoversikt-skrivestotte.Application")
+
+fun main() {
     val configuration = Configuration()
     val dbConfig = DataSourceConfiguration(configuration)
-
     DataSourceConfiguration.migrateDb(dbConfig.adminDataSource())
 
     HttpServer.create("modiapersonoversikt-skrivestotte", 7070) {
         skrivestotteApp(
             configuration = configuration,
-            dataSource = dbConfig.userDataSource(),
-            useAuthentication = useAuthentication
+            dataSource = dbConfig.userDataSource()
         )
     }.start(wait = true)
-}
-
-fun main() {
-    runLocally(true)
 }
