--- src/test/kotlin/no/nav/modiapersonoversikt/frequencies.txt
@@ -1,139 +1,0 @@
-7092
-1738
-1663
-1605
-998
-981
-884
-776
-656
-440
-439
-389
-339
-332
-196
-247
-233
-228
-226
-210
-203
-191
-186
-180
-176
-166
-164
-156
-154
-146
-146
-141
-139
-135
-132
-122
-103
-102
-99
-99
-98
-94
-87
-75
-70
-66
-66
-63
-61
-58
-55
-54
-54
-51
-49
-48
-45
-43
-43
-42
-42
-41
-39
-38
-37
-36
-35
-35
-35
-33
-33
-33
-33
-33
-31
-31
-30
-29
-28
-28
-27
-26
-25
-25
-25
-24
-24
-23
-22
-21
-20
-20
-20
-20
-19
-18
-18
-18
-17
-17
-17
-17
-17
-17
-16
-16
-16
-16
-16
-16
-15
-15
-14
-14
-14
-13
-13
-13
-13
-13
-13
-13
-13
-12
-12
-12
-12
-12
-12
-11
-11
-11
-11
-11
-11
-11
-11
-11
-10
